# apps/miniapp/src/services/api/types.ts
import type { components } from '../../../../../packages/api-client/src/generated'
export type Member = components['schemas']['MemberDto']
export type Session = components['schemas']['SessionDto']
export type Activity = components['schemas']['ActivityDto']
export type Registration = components['schemas']['RegistrationDto']
export type MyRegistration = components['schemas']['MyRegistrationDto']
export type RegistrationInput = components['schemas']['RegistrationInput']
export type PublicMember = components['schemas']['PublicMemberDto']
export type ActivityWriteInput = components['schemas']['ActivityWriteInput']
export type ManagedActivity = components['schemas']['ManagedActivityDto']
export type OrganizerRegistration = components['schemas']['OrganizerRegistrationDto']
export type ActivityNotification = components['schemas']['NotificationDto']
export type ReadingList = components['schemas']['ReadingListDto']
export type ReadingStats = components['schemas']['ReadingStatsDto']

export type ProfileDetails = components['schemas']['ProfileDetailsDto']
export type ProfileDetailsInput = components['schemas']['ProfileDetailsInput']
export type CardSettings = components['schemas']['CardSettingsDto']
export type MemberCard = components['schemas']['MemberCardDto']